requirements.txt
Flask
gunicorn
python-dotenv
google-generativeai
pandas
numpy